cooldownRemainingText refactor
