+++ src/commands/attack.ts
@@ -4,7 +4,8 @@ import { getUserCharacter } from "../character/getUserCharacter";
 import { getCharacterStatModifier } from "../character/getCharacterStatModifier";
 import { getCharacterStatModified } from "../character/getCharacterStatModified";
 import { playerAttack } from "../attack/playerAttack";
-import { cooldownRemainingText, sleep } from "../utils";
+import { sleep } from "../utils";
+import { cooldownRemainingText } from "../character/cooldownRemainingText";
 import { mentionCharacter } from "../character/mentionCharacter";
 import { attack } from "../attack/attack";
 import { hpBarField } from "../character/hpBar/hpBarField";
