+++ src/commands/heal.ts
@@ -2,7 +2,7 @@ import { SlashCommandBuilder } from "@discordjs/builders";
 import { CommandInteraction, Message, MessageEmbed } from "discord.js";
 import { heal } from "../heal/heal";
 import { getUserCharacter } from "../character/getUserCharacter";
-import { cooldownRemainingText } from "../utils";
+import { cooldownRemainingText } from "../character/cooldownRemainingText";
 import { hpBarField } from "../character/hpBar/hpBarField";
 import { Emoji } from "../Emoji";
 import { updateUserQuestProgess } from "../quest/updateQuestProgess";
